Refactoring user resources
