+++ src/Resources/User/Create.ts
@@ -1,0 +1,13 @@
+import AbstractResource from '../AbstractResource';
+
+export default class Create extends AbstractResource {
+    initialise(): void {
+        this.resourcePath = '/users';
+        this.defaultAccess = 'guest';
+    }
+
+    public create: Function = this.createMethodFromPartialSpec({
+        method: 'POST',
+        path: '/',
+    });
+}
