+++ src/Resources/User/Details.ts
@@ -1,18 +1,12 @@
 import AbstractResource from '../AbstractResource';
 
-export default class Users extends AbstractResource {
+export default class UserDetails extends AbstractResource {
     initialise(): void {
         this.resourcePath = '/users';
-        this.defaultAccess = 'guest';
+        this.defaultAccess = 'member';
     }
 
-    public create: Function = this.createMethodFromPartialSpec({
-        method: 'POST',
-        path: '/',
-    });
-
     public update: Function = this.createMethodFromPartialSpec({
-        access: 'member',
         method: 'POST',
         path: '/{id}',
     });
