+++ src/Resources/User/Subscriptions.ts
@@ -1,4 +1,4 @@
-import AbstractResource from '../../AbstractResource';
+import AbstractResource from '../AbstractResource';
 
 export default class Subscriptions extends AbstractResource {
     initialise(): void {
