+++ src/Resources/User/Toybox.ts
@@ -1,4 +1,4 @@
-import AbstractResource from '../../AbstractResource';
+import AbstractResource from '../AbstractResource';
 
 export default class Toybox extends AbstractResource {
     initialise(): void {
