+++ src/Resources/User/index.ts
@@ -1,18 +1,22 @@
 import Client from '../../Client';
-import Basket from './Basket/Basket';
-import Subscriptions from './Subscriptions/Subscriptions';
-import Toybox from './Toybox/Toybox';
+import Basket from './Basket';
+import Subscriptions from './Subscriptions';
+import Toybox from './Toybox';
+import Details from './Details';
+import Create from './Create';
 
-export interface UserResources {
-    basket: Basket;
-    subscriptions: Subscriptions;
-    toybox: Toybox;
-}
+export default class User {
+    public basket: Basket;
+    public subscriptions: Subscriptions;
+    public toybox: Toybox;
+    public details: Details;
+    public create: Create;
 
-export function loadUserResources(client: Client): UserResources {
-    return {
-        basket: new Basket(client),
-        subscriptions: new Subscriptions(client),
-        toybox: new Toybox(client),
-    };
+    constructor(client: Client) {
+        this.basket = new Basket(client);
+        this.subscriptions = new Subscriptions(client);
+        this.toybox = new Toybox(client);
+        this.details = new Details(client);
+        this.create = new Create(client);
+    }
 }
