+++ src/__tests__/Client.spec.ts
@@ -1,7 +1,6 @@
 import { httpClient } from '../../test/mocks/HttpClient';
 import Client from '../Client';
 import Auth from '../Resources/Auth/Auth';
-import Users from '../Resources/Users/Users';
 import WACCOrders from '../Resources/Wacc/Orders/Orders';
 import WACCReturnOrders from '../Resources/Wacc/ReturnOrders/ReturnOrders';
 import WACCSearch from '../Resources/Wacc/Search/Search';
@@ -9,11 +8,12 @@ import WACCUsers from '../Resources/Wacc/Users/Users';
 import Subscriptions from '../Resources/Subscriptions/Subscriptions';
 import SubscriptionAddons from '../Resources/Subscriptions/SubscriptionAddons/SubscriptionAddons';
 import Codes from '../Resources/Codes/Codes';
-import Toybox from '../Resources/User/Toybox/Toybox';
+import Toybox from '../Resources/User/Toybox';
 import Products from '../Resources/Products/Products';
 import GuestBasket from '../Resources/Guest/Basket/Basket';
-import UserBasket from '../Resources/User/Basket/Basket';
-import UserSubscriptions from '../Resources/User/Subscriptions/Subscriptions';
+import User from '../Resources/User';
+import UserBasket from '../Resources/User/Basket';
+import UserSubscriptions from '../Resources/User/Subscriptions';
 
 describe('Client.ts', () => {
     let client: Client;
@@ -33,7 +33,6 @@ describe('Client.ts', () => {
     it('can access all resources', () => {
         // @todo We could probably turn this in to a loop to make it easier
         expect(client.auth).toBeInstanceOf(Auth);
-        expect(client.users).toBeInstanceOf(Users);
         expect(client.subscriptions.subscriptions).toBeInstanceOf(Subscriptions);
         expect(client.subscriptions.addons).toBeInstanceOf(SubscriptionAddons);
         expect(client.codes).toBeInstanceOf(Codes);
@@ -43,6 +42,7 @@ describe('Client.ts', () => {
         expect(client.wacc.search).toBeInstanceOf(WACCSearch);
         expect(client.wacc.users).toBeInstanceOf(WACCUsers);
         expect(client.guest.basket).toBeInstanceOf(GuestBasket);
+        expect(client.user).toBeInstanceOf(User);
         expect(client.user.basket).toBeInstanceOf(UserBasket);
         expect(client.user.subscriptions).toBeInstanceOf(UserSubscriptions);
         expect(client.user.toybox).toBeInstanceOf(Toybox);
